Default to highp

+++ src/goo/renderer/Renderer.js
@@ -207,7 +207,7 @@ function (
 			fragmentShaderLowpInt: this.context.getShaderPrecisionFormat(this.context.FRAGMENT_SHADER, this.context.LOW_INT)
 		};
 
-		this.shaderPrecision = parameters.shaderPrecision || 'mediump';
+		this.shaderPrecision = parameters.shaderPrecision || 'highp';
 		if (this.shaderPrecision === 'highp' && this.capabilities.vertexShaderHighpFloat.precision > 0 && this.capabilities.fragmentShaderHighpFloat.precision > 0) {
 			this.shaderPrecision = 'highp';
 		} else if (this.shaderPrecision !== 'lowp' && this.capabilities.vertexShaderMediumpFloat.precision > 0 && this.capabilities.fragmentShaderMediumpFloat.precision > 0) {
